requirements.txt
# Reddit API Client Dependencies

# Reddit API wrapper
praw==7.7.1

# Environment variable management
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
